+++ example/demo-ui-tools.js
@@ -15,7 +15,7 @@ export const uiLoggedInState = (loggedIn) => {
   const loginExtensionButton = document.getElementById(
     'button-login-extension'
   );
-  const loginMaiarButton = document.getElementById('button-login-mobile');
+  const loginButton = document.getElementById('button-login-mobile');
   const loginWebButton = document.getElementById('button-login-web');
   const logoutButton = document.getElementById('button-logout');
   const txButton = document.getElementById('button-tx');
@@ -24,7 +24,7 @@ export const uiLoggedInState = (loggedIn) => {
   const queryButton = document.getElementById('button-query');
   if (loggedIn) {
     loginExtensionButton.style.display = 'none';
-    loginMaiarButton.style.display = 'none';
+    loginButton.style.display = 'none';
     loginWebButton.style.display = 'none';
     logoutButton.style.display = 'block';
     txButton.style.display = 'block';
@@ -33,7 +33,7 @@ export const uiLoggedInState = (loggedIn) => {
     queryButton.style.display = 'block';
   } else {
     loginExtensionButton.style.display = 'block';
-    loginMaiarButton.style.display = 'block';
+    loginButton.style.display = 'block';
     loginWebButton.style.display = 'block';
     logoutButton.style.display = 'none';
     txButton.style.display = 'none';
@@ -48,7 +48,7 @@ export const updateTxHashContainer = (txHash) => {
   const txHashContainer = document.getElementById('tx-hash-or-query-result');
   if (txHash) {
     txHashContainer?.replaceChildren();
-    const url = `https://devnet-explorer.elrond.com/transactions/${txHash}`;
+    const url = `https://devnet-explorer.multiversx.com/transactions/${txHash}`;
     const link = document.createElement('a');
     link.setAttribute('href', url);
     link.classList.add('transaction-link');