+++ src/auth/init-web-wallet-provider.ts
@@ -1,4 +1,4 @@
-import { WalletProvider } from '@elrondnetwork/erdjs-web-wallet-provider/out/walletProvider';
+import { WalletProvider } from '@multiversx/sdk-web-wallet-provider/out/walletProvider';
 import { ls } from '../utils/ls-helpers';
 import { getParamFromUrl } from '../utils/get-param-from-url';
 import { DAPP_INIT_ROUTE } from '../utils/constants';