+++ src/auth/login-with-web-wallet.ts
@@ -1,4 +1,4 @@
-import { WalletProvider } from '@elrondnetwork/erdjs-web-wallet-provider/out/walletProvider';
+import { WalletProvider } from '@multiversx/sdk-web-wallet-provider/out/walletProvider';
 import { EventStoreEvents, LoginMethodsEnum } from '../types';
 import { DAPP_INIT_ROUTE } from '../utils/constants';
 import { errorParse } from '../utils/error-parse';