+++ src/main.ts
@@ -1,9 +1,9 @@
-import { Transaction } from '@elrondnetwork/erdjs/out/transaction';
+import { Transaction } from '@multiversx/sdk-core/out/transaction';
 import { initExtensionProvider } from './auth/init-extension-provider';
-import { ExtensionProvider } from '@elrondnetwork/erdjs-extension-provider/out/extensionProvider';
-import { WalletConnectV2Provider } from '@elrondnetwork/erdjs-wallet-connect-provider/out/walletConnectV2Provider';
-import { WalletProvider } from '@elrondnetwork/erdjs-web-wallet-provider/out/walletProvider';
-import { initMaiarMobileProvider } from './auth/init-maiar-mobile-provider';
+import { ExtensionProvider } from '@multiversx/sdk-extension-provider/out/extensionProvider';
+import { WalletConnectV2Provider } from '@multiversx/sdk-wallet-connect-provider/out/walletConnectV2Provider';
+import { WalletProvider } from '@multiversx/sdk-web-wallet-provider/out/walletProvider';
+import { initMobileProvider } from './auth/init-mobile-provider';
 import { ls } from './utils/ls-helpers';
 import { ApiNetworkProvider, SmartContractQueryArgs } from './network-provider';
 import {
@@ -108,11 +108,11 @@ export class ElvenJS {
     if (isAddress && state?.loginMethod) {
       EventsStore.run(EventStoreEvents.onLoginPending);
 
-      if (state.loginMethod === LoginMethodsEnum.maiarBrowserExtension) {
+      if (state.loginMethod === LoginMethodsEnum.browserExtension) {
         this.dappProvider = await initExtensionProvider();
       }
-      if (state.loginMethod === LoginMethodsEnum.maiarMobile) {
-        this.dappProvider = await initMaiarMobileProvider(this);
+      if (state.loginMethod === LoginMethodsEnum.mobile) {
+        this.dappProvider = await initMobileProvider(this);
       }
       if (
         state.loginMethod === LoginMethodsEnum.webWallet &&
@@ -155,14 +155,14 @@ export class ElvenJS {
     }
 
     try {
-      // Login with Maiar browser extension
-      if (loginMethod === LoginMethodsEnum.maiarBrowserExtension) {
+      // Login with browser extension
+      if (loginMethod === LoginMethodsEnum.browserExtension) {
         const dappProvider = await loginWithExtension(this, options?.token);
         this.dappProvider = dappProvider;
       }
 
-      // Login with Maiar mobile app
-      if (loginMethod === LoginMethodsEnum.maiarMobile) {
+      // Login with mobile app
+      if (loginMethod === LoginMethodsEnum.mobile) {
         const dappProvider = await loginWithMobile(
           this,
           options?.qrCodeContainer,